chore: Add SampSharp.CTF.Entities package
The SampSharp.CTF.Entities package is similar to the SampSharp.Entities package, what makes it different is that the former has the latest updates of SampSharp.Entities.

The author has not released a new version of SampSharp.Entities, so I had to recompile the project.

See https://www.nuget.org/packages/SampSharp.CTF.Entities