fix: Remove private field called `_connectedPlayer`
This fixes a bug: If a player connects to the server, it overwrites the current value of "_connectedPlayer", this is a bad idea if two players are trying to log in or register at the same time.

Note also that the service called “AccountSystem” is added as a singleton by SampSharp.