BREAKING CHANGE: Remove EntityId property from PlayerInfo class
It is not necessary to couple the "PlayerInfo" class with the "EntityId" type of SampSharp.

The disadvantage of having an "EntityId" property, is that it enters in confusion with the AccountId property, at first sight the consumer does not know what is the difference between them, unless he checks the XML comments.