BREAKING CHANGE: Inject an argument of type TimeLeft instead of a string (#92)
This change is implemented to ensure that the "UpdateTimeLeft" method only updates the text in a fixed format. Previously, a string could be passed in any format, which could lead to inconsistencies or errors in how the time was displayed.

By using the "TimeLeft" class, we ensure that the text format is always "00:00", as required for correct time display. The "TimeLeft" class enforces this format, preventing any incorrect formats from being used and improving the consistency and reliability of the time presentation.