fix(sqlite): Remove 'REGEXP' operator from CHECK constraint (#200)
The 'REGEXP' operator is causing this exception when creating a player or updating the player name:
'System.AccessViolationException' occurred in SQLitePCLRaw.provider.e_sqlite3.dll Attempted to read or write protected memory. This is often an indication that other memory is corrupt.

The strange thing is that this exception only occurs when using the 32-bit e_sqlite.dll.